fix:服务修改

+++ server/service/ShellService.js
@@ -1,9 +1,16 @@
+import { Console } from 'console';
+import { cpuUsage } from 'process';
+
 const spawn = require('cross-spawn');
 const fs = require('fs');
 const { execFile } = require('child_process');
 const path = require('path');
-const { clear } = require('console');
 const process = require('process');
+
+// 解决windows命令行编码和node默认编码不同导致的中文乱码
+var iconv = require('iconv-lite');
+var encoding = 'cp936';
+var binaryEncoding = 'binary';
 
 const workPath = process.cwd(); // node.js进程当前工作目录
 
@@ -13,13 +20,21 @@ class ShellService {
     }
 
     runShell(project, logDir, cmdDir, moduleBaseDir, webSocket) {
+        let response = {
+            state: '', // 标志当前进程的状态 1表示运行中， 0 表示运行结束
+            rst: '', // 标志当前进行执行结果, 1表示运行成功， 0表示运行失败
+            data: '', // 指令执行过程中的消息
+        };
+
         // 如果当前运行工程列表中已经存在该项目则返回
         const findProjectIndex = this.runProjectList.findIndex((item) => {
             return item.projectId === project.projectId;
         });
 
         if (findProjectIndex > -1) {
-            return;
+            const findProject = this.runProjectList[findProjectIndex];
+            findProject.process.kill('SIGINT');
+            this.runProjectList.splice(findProjectIndex, 1);
         }
 
         // 日志路径
@@ -46,7 +61,7 @@ class ShellService {
         // 如果当前工程配置了需要打包的模块数据，则将模块信息写入package.json
         if (fs.existsSync(modulePath) && project.moduleChoice) {
             if (!fs.existsSync(packagePath)) {
-                throw new Error('package.json不存在');
+                throw new Error('当前工程下不存在package.json文件');
             }
             let data = fs.readFileSync(modulePath).toString();
             let dataObj = [];
@@ -60,7 +75,7 @@ class ShellService {
             fs.writeFileSync(packagePath, JSON.stringify(packageData, null, 4));
         } else if (project.moduleChoice) {
             // 说明
-            throw new Error('项目的不存在模块配置');
+            throw new Error('项目的不存在模块配置 ');
         }
 
         // 运行工程如果有配置路径，则先将'导航到指定的路径下'的命令拼接到命令集中
@@ -69,7 +84,23 @@ class ShellService {
 
         fs.appendFileSync(cmdPath, cmdSet);
 
-        const process = execFile(cmdPath, [], { maxBuffer: 5000 * 1024 });
+        const process = execFile(
+            cmdPath,
+            [],
+            {
+                maxBuffer: 5000 * 1024,
+                encoding: binaryEncoding,
+            },
+            (error, stdout, stderr) => {
+                if (error) {
+                    response = {
+                        state: 1,
+                        data: error,
+                    };
+                }
+                webSocket.send(JSON.stringify(response));
+            }
+        );
 
         const runInfo = {
             projectName: project.projectName,
@@ -81,44 +112,92 @@ class ShellService {
         this.runProjectList.push(runInfo);
 
         process.stdout.on('data', (data) => {
-            fs.appendFileSync(logPath, data);
-            webSocket.send(data);
+            const decodeData = iconv.decode(new Buffer(data, binaryEncoding), encoding);
+            fs.appendFileSync(logPath, decodeData);
+
+            response = {
+                state: 1,
+                data: decodeData,
+            };
+
+            webSocket.send(JSON.stringify(response));
         });
 
         process.stdout.on('error', (data) => {
-            fs.appendFileSync(logPath, data);
-            // 当命令发送错误
-            // 就将其从this.runProjectList中移除
+            const decodeData = iconv.decode(new Buffer(data, binaryEncoding), encoding);
+            fs.appendFileSync(logPath, decodeData);
+
+            response = {
+                state: 0,
+                data: decodeData,
+                method: 'run',
+            };
+
+            webSocket.send(JSON.stringify(response));
+        });
+
+        process.stderr.on('data', (data) => {
+            const decodeData = iconv.decode(new Buffer(data, binaryEncoding), encoding);
+            fs.appendFileSync(logPath, decodeData);
+
+            response = {
+                state: 1,
+                data: decodeData,
+            };
+
+            webSocket.send(JSON.stringify(response));
+        });
+
+        process.stderr.on('error', (data) => {
+            const decodeData = iconv.decode(new Buffer(data, binaryEncoding), encoding);
+            fs.appendFileSync(logPath, decodeData);
+
+            response = {
+                state: 0,
+                data: decodeData,
+                method: 'run',
+            };
+
+            webSocket.send(JSON.stringify(response));
+        });
+
+        // process.on('exit', (code, signal) => {
+        //     console.log('退出码', code);
+        //     fs.appendFileSync(logPath, `子进程退出，退出码 ${code}`);
+
+        //     // 就将其从this.runProjectList中移除
+        //     const findProjectIndex = this.runProjectList.findIndex((item) => {
+        //         return item.projectId === project.projectId;
+        //     });
+
+        //     response = {
+        //         state: 1,
+        //         data: '',
+        //         rst: code,
+        //     };
+
+        //     webSocket.send(JSON.stringify(response));
+        // });
+
+        process.on('close', (code, signal) => {
             const findProjectIndex = this.runProjectList.findIndex((item) => {
                 return item.projectId === project.projectId;
             });
+
             this.runProjectList.splice(findProjectIndex, 1);
 
-            webSocket.send(data);
-        });
+            response = {
+                state: 0,
+                data: '运行结束',
+                rst: code,
+                method: 'run',
+            };
 
-        process.on('exit', (code, signal) => {
-            fs.appendFileSync(logPath, `子进程退出，退出码 ${code}`);
-
-            webSocket.send(signal);
-        });
-
-        process.on('close', (code) => {
-            fs.appendFileSync(logPath, `子进程关闭，退出码 ${code}`);
-            // 当命令运行完成之后，当命令运行完毕之后，需要进行一些清理工作
-            // 就将其从this.runProjectList中移除
-            const findProjectIndex = this.runProjectList.findIndex((item) => {
-                return item.projectId === project.projectId;
-            });
-            this.runProjectList.splice(findProjectIndex, 1);
-            // 清空当前工程的bat文件内容
-            // fs.writeFileSync(cmdPath, '');
-
-            webSocket.send(code);
+            webSocket.send(JSON.stringify(response));
         });
     }
 
-    suspendRun(projectId) {
+    suspendRun(projectId, webSocket) {
         let findProjectIndex = -1;
         const findProject = this.runProjectList.find((item, index) => {
             if (item.projectId === projectId) {
@@ -127,28 +206,19 @@ class ShellService {
             return item;
         });
 
-        findProject.process.kill('SIGINT');
-        this.runProjectList.splice(findProjectIndex, 1);
-    }
-
-    runLog(projectId, logDir, callback) {
-        const logPath = `${logDir}/${projectId}.txt`;
-
-        const data = fs.readFileSync(logPath);
-
-        const findProjectIndex = this.runProjectList.findIndex((item) => {
-            return item.projectId === projectId;
-        });
-        if (findProjectIndex < 0) {
-            return {
-                status: 0,
-                log: data.toString(),
-            };
+        if (findProject) {
+            findProject.process.kill('SIGINT');
+            this.runProjectList.splice(findProjectIndex, 1);
         }
-        return {
-            status: 1,
-            log: data.toString(),
+
+        let response = {
+            state: 0, //
+            rst: '',
+            data: '项目已中止',
+            method: 'suspend',
         };
+
+        webSocket.send(JSON.stringify(response));
     }
 }
